Create python-app.yml
Permissions are needed to update the Actions interface.  Permissions can be added under the `job` section for a specific job, or at the top of a workflow so all jobs assume the permission.
Tests can be updated to include JUnit reports.
Actions can be added to publish JUnit reports to the Actions user interface.